moved get_models out of daq_utils and into utils in the pid package
