+++ src/pymodaq/extensions/__init__.py
@@ -4,9 +4,13 @@ Created the 25/10/2022
 
 @author: Sebastien Weber
 """
+from .utils import get_extensions
+from .pid.utils import get_models
 
 from .console import QtConsole
 from .daq_scan import DAQ_Scan
 from .daq_logger import DAQ_Logger
 from .pid.pid_controller import DAQ_PID
 from .h5browser import H5Browser
+
+
